+++ app/(root)/terms-and-conditions.tsx
@@ -1,0 +1,5 @@
+import { TermsAndConditionsScreen } from 'appdeptus/modules/root/screens'
+
+const TermsAndConditionsRoute = () => <TermsAndConditionsScreen />
+
+export default TermsAndConditionsRoute