+++ modules/root/components/PrivacyLink/PrivacyLink.tsx
@@ -1,9 +1,0 @@
-import { Text, TextLink } from 'appdeptus/components'
-
-const PrivacyLink = () => (
-  <Text>
-    I accept the <TextLink href='privacy-policy'>privacy policy</TextLink>
-  </Text>
-)
-
-export default PrivacyLink