+++ modules/root/components/TermsAndConditions/TermsAndConditions.tsx
@@ -1,0 +1,10 @@
+import { Text, TextLink } from 'appdeptus/components'
+
+const TermsAndConditions = () => (
+  <Text>
+    I accept the{' '}
+    <TextLink href='terms-and-conditions'>terms and conditions</TextLink>
+  </Text>
+)
+
+export default TermsAndConditions